failing, functional odometry unitTest
Unit test works, but is not passing.
Well, there are no assertions at this point, but 
they would fail if they existed.

Added FakeTelemetry to print more runtime debug 
outputs in the unit test.

Revised DeadWheelOdometry slightly to make 
testing easier.
odometryLoop() was renamed odometryCalculation, 
and now the currentEncoderValues are passed in, 
rather than being grabbed directly.  This allows this 
function to be spoofed easily in a test.